requirements.txt
python-telegram-bot[ext]==21.3
openai>=1.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
